chore(pointcloud): move computeVisibility method to utils
